Revert "Add search_files.py, refactor common functionality into common.py"
