+++ README.md
@@ -68,63 +68,3 @@ options:
                         complete with an interval of this many minutes. Non-
                         positive values will disable monitoring.
 ```
-
-## search_files
-
-Performs DataGateway searches for Datafiles matching the provided query.
-```bash
-python3 search_files.py input-file.txt 'visitId:"AB1234-1"' --username=abc12345
-```
-
-Full help text and description of arguments is available with the `--help` command:
-```bash
-python3 search_files.py --help
-```
-```
-usage: search_files [-h] [--url URL] [-a AUTHENTICATOR] -u USERNAME
-                    [-p PASSWORD_FILE] [-m MAX_RESULTS]
-                    output_file query
-
-Performs DataGateway searches for Datafiles matching the provided query. These paths will be written to file in
-batches, and can then be inspected and filtered further if needed before submitting using queue_file_downloads.
-
-positional arguments:
-  output_file           File to append newline separated paths to.
-                        This can then be provided as an input to queue_file_downloads.
-  query                 Lucene syntax formatted search query. Full help text and examples can be found
-                        in the DataGateway UI. Note that wildcards can significantly increase the time
-                        taken to perform a search, and the more specific the search query is the more
-                        efficient it will be. Some example searches are:
-                            'visitId:AB1234'
-                                Search for all Datafiles in all parts of proposal
-                            'visitId:"AB1234-1"'
-                                Search for all Datafiles in a (part) visit
-                            'location.fileName:"config.txt"'
-                                Search for Datafiles with a specific file name and extension (both required)
-                            'location.fileName:config'
-                                Search for Datafiles with a specific name but any extension
-                            'location.fileName:txt'
-                                Search for Datafiles with the extension 'txt', but no requirement on the name
-                            'location:raw'
-                                Search for Datafiles with the directory 'raw' somewhere in their path
-                            'location:(raw processed)'
-                                Search for Datafiles with either of two directories somewhere in their path
-                            'location.exact:/dls/i0/data/2000'
-                                Search for Datafiles in any subdirectory of the provided path (case sensitive)
-                            'location.exact:/dls/i0/data/202?/*/raw/config.txt'
-                                Search for a full path with wildcards (case sensitive)
-                            '+location.exact:/dls/i0/data/202? +location:(raw processed) +location.fileName:txt'
-                                Search for multiple criteria (all of which are required to match)
-
-optional arguments:
-  -h, --help            show this help message and exit
-  --url URL             The url address of the DataGateway instance to submit requests to.
-  -a AUTHENTICATOR, --authenticator AUTHENTICATOR
-                        The authentication mechanism to use for DataGateway login.
-  -u USERNAME, --username USERNAME
-                        The username used for DataGateway login.
-  -p PASSWORD_FILE, --password-file PASSWORD_FILE
-                        Location of file containing password for DataGateway login. If not provided, the password will need to be provided by prompt.
-  -m MAX_RESULTS, --max-results MAX_RESULTS
-                        The maximum number of results to request in a single batch. If unset, the server default value will be used.
-```
